pyproject.toml
[project]
name = "first-agent"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9"
dependencies = [
    "chromadb>=1.4.0",
    "google-genai>=1.47.0",
    "openai>=2.15.0",
    "playwright>=1.57.0",
    "psutil>=7.2.1",
    "pydantic-ai>=0.8.1",
    "python-dotenv>=1.2.1",
    "schedule>=1.2.2",
]
